Edit report : add sciprt_id parameter to performance report

+++ services/report/src/main/java/org/ats/service/report/jmeter/Calculation.java
@@ -51,7 +51,7 @@ public class Calculation {
       if (point == null) {
         point = new PointReport();
         point.setTimestamp(bucketHis);
-        point.setDate(new Date(bucketHis));
+        point.setDate(new Date(bucketHis.longValue()));
         point.setValue(0);
       }
       point.setValue(point.getValue() + 1);
@@ -66,7 +66,7 @@ public class Calculation {
       if (point == null) {
         point = new PointReport();
         point.setTimestamp(bucketTrans);
-        point.setDate(new Date(bucketTrans));
+        point.setDate(new Date(bucketTrans.longValue()));
         point.setValue(0);
       }
       point.setValue(point.getValue() + 1);
@@ -94,7 +94,7 @@ public class Calculation {
     summary.setKbPerSecond(summary.totalBytes * MILLIS_PER_SECOND / KB_TO_BY / (summary.lastTimeFinal - summary.firsTimestamp));
 
     // averageBytes
-    summary.setAverageBytes(summary.totalBytes / summary.samples);
+    summary.setAverageBytes((double)summary.totalBytes / summary.samples);
   }
 
   private static double calculateStDev(int number, int elementN1, double prev_avg, double pre_std_dev, double current_avg) {
